requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
google-generativeai==0.5.1   # Gemini SDK
chromadb==0.4.24
sentence-transformers==3.0.0
pymupdf==1.24.4
python-docx==1.2.1
mail-parser==3.15.0
python-dotenv==1.0.1
pydantic==2.7.1
sqlalchemy==2.0.32
pytest==8.2.1
httpx==0.27.0
